Synth Project- Playing Chords on a Sine Wave ~~~
Watched another couple YouTube tutorials and downloaded some more repositories. Need to incorporate the waves I was building as they sound a bit cleaner. Borrowed chords from a song called Borderline by Tame Impala, messed with the duration of the notes until it sounded nicer. Doesn't really sound anything like the song because of how much rhythm he adds to it. and the sound of the instrument is a lot cooler.